fixed bugs, cleaned and rearranged code
