+++ backend/src/main/java/com/authenticket/authenticket/config/SecurityConfig.java
@@ -67,7 +67,8 @@ public class SecurityConfig {
                         .requestMatchers("/api/order/**").hasAuthority("USER")
 
                         .requestMatchers("/api/ticket-category/**").hasAuthority("ADMIN")
-                        .requestMatchers("/api/ticket/**").hasAuthority("ADMIN")
+
+                        .requestMatchers("/api/ticket/**").hasAuthority("USER")
 
                         .requestMatchers(HttpMethod.PUT, "/api/user/updateUserProfile").hasAuthority("USER")
                         .requestMatchers(HttpMethod.PUT, "/api/user/updateUserImage").hasAuthority("USER")
