+++ backend/src/main/java/com/authenticket/authenticket/controller/TicketController.java
@@ -2,12 +2,17 @@ package com.authenticket.authenticket.controller;
 
 import com.authenticket.authenticket.dto.ticket.TicketDisplayDto;
 import com.authenticket.authenticket.model.Ticket;
+import com.authenticket.authenticket.repository.TicketRepository;
+import com.authenticket.authenticket.service.Utility;
 import com.authenticket.authenticket.service.impl.TicketServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.webjars.NotFoundException;
 
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @RestController
 @CrossOrigin(
@@ -21,16 +26,24 @@ import java.util.List;
         allowCredentials = "true"
 )
 @RequestMapping("/api/ticket")
-public class TicketController {
+public class TicketController extends Utility {
     private final TicketServiceImpl ticketService;
+    private final TicketRepository ticketRepository;
     @Autowired
-    public TicketController(TicketServiceImpl ticketService) {
+    public TicketController(TicketServiceImpl ticketService, TicketRepository ticketRepository) {
         this.ticketService = ticketService;
+    this.ticketRepository = ticketRepository;
     }
 
     @GetMapping("/test")
     public String test() {
         return "test successful";
+    }
+
+    @PostMapping("/test-post")
+    public String testPost(@RequestParam(value = "eventId") Integer eventId,
+                           @RequestParam(value = "sectionId") Integer sectionId) {
+        return ticketRepository.findNoOfAvailableTicketsBySectionAndEvent(eventId,sectionId).toString();
     }
 
     @GetMapping
@@ -41,6 +54,32 @@ public class TicketController {
     @GetMapping("/{ticketId}")
     public ResponseEntity<?> findTicketById(@PathVariable("ticketId") Integer ticketId) {
         return ResponseEntity.ok(ticketService.findTicketById(ticketId));
+    }
+
+    @PostMapping("/allocate-seats")
+    public ResponseEntity<?> allocateSeats(@RequestParam(value = "eventId") Integer eventId,
+                                           @RequestParam(value = "sectionId") Integer sectionId,
+                                           @RequestParam(value = "ticketsToPurchase") Integer ticketsToPurchase) {
+        List<Ticket> ticketList;
+        try{
+            ticketList = ticketService.allocateSeats(eventId, sectionId, ticketsToPurchase);
+        } catch(Exception e){
+            return ResponseEntity.badRequest().body(generateApiResponse(null, e.getMessage()));
+        }
+        return ResponseEntity.ok(generateApiResponse(ticketList, String.format("%d seats successfully assigned", ticketsToPurchase)));
+    }
+
+    @DeleteMapping("/unallocate-seats")
+    public ResponseEntity<?> removeTickets(@RequestParam(value = "ticketIdString")String ticketIdString
+    ) {
+        List<Integer> ticketIdList = Arrays.stream(ticketIdString.split(","))
+                .map(Integer::parseInt)
+                .collect(Collectors.toList());
+
+
+            ticketService.removeAllTickets(ticketIdList);
+
+        return ResponseEntity.ok(generateApiResponse(null, "Tickets Unallocated Successfully"));
     }
 
 //    @PostMapping
