+++ backend/src/main/java/com/authenticket/authenticket/repository/TicketRepository.java
@@ -16,8 +16,14 @@ public interface TicketRepository extends JpaRepository<Ticket, Integer> {
     List<Ticket> findAllByEventEventIdAndSectionSectionId(Integer eventId, Integer sectionId);
     List<Ticket> findAllByEventEventIdAndSectionSectionIdAndRowNo(Integer eventId, Integer sectionId,Integer rowNo);
     Integer countByEventEventIdAndSectionSectionIdAndRowNo(Integer eventId, Integer sectionId, Integer rowNo);
-    Integer countByEventEventIdAndSectionSectionId(Integer eventId, Integer sectionId);
 
+    @Query(
+            nativeQuery = true, value="SELECT " +
+            "(s.no_of_rows * s.no_of_seats_per_row - CAST(COUNT(t) AS INTEGER)) " +
+            "FROM dev.section s " +
+            "JOIN dev.ticket t ON s.section_id = t.section_id AND t.event_id = :eventId AND t.section_id = :sectionId " +
+            "GROUP BY s.section_id, s.no_of_rows, s.no_of_seats_per_row, s.category_id")
+    Integer findNoOfAvailableTicketsBySectionAndEvent(@Param("eventId") Integer eventId, @Param("sectionId") Integer sectionId);
     @Query(
             nativeQuery = true, value="SELECT " +
             "s.section_id, " +
